// komg/komgo-member-api-trade-cargo/src/business-layer/messaging/enums.ts
export enum PRODUCT_ID {
  TradeFinance = 'tradeFinance'
}

export enum NOTIFICATION_TYPE {
  LCTask = 'LC.task',
  LCInfo = 'LC.info'
}

export enum NOTIFICATION_USER {
  ManageTrades = 'manageTrades',
  ManageCargo = 'manageCargo'
}

export enum NOTIFICATION_LEVEL {
  info = 'info',
  warning = 'warning',
  danger = 'danger'
}